app/src/main/java/com/codesoom/assignment/application/ProductService.java
package com.codesoom.assignment.application;

import com.codesoom.assignment.domain.Product;
import com.codesoom.assignment.domain.ProductRepository;
import org.springframework.stereotype.Service;
import org.springframework.transaction.annotation.Transactional;

import java.util.List;


/** TODO
 1. get all product
 2. get product by id
 3. create product
 4. update product by id
 5. delete product by id
*/
@Service
@Transactional
public class ProductService {

    private final ProductRepository productRepository;

    public ProductService(ProductRepository productRepository) {
        this.productRepository = productRepository;
    }

    @Transactional(readOnly = true)
    public List<Product> getList() {
        return productRepository.findAll();
    }
}
